fix(movement): split motion controls for module budget
