+++ src/movement_optimizer/gui/motion_tabs.py
@@ -7,7 +7,7 @@ from collections.abc import Callable, Sequence
 from itertools import pairwise
 
 import numpy as np
-from PyQt6.QtCore import QPointF, Qt, QTimer, pyqtSignal
+from PyQt6.QtCore import QPointF, QTimer, pyqtSignal
 from PyQt6.QtGui import QColor, QPainter, QPen
 from PyQt6.QtWidgets import (
     QCheckBox,
@@ -16,13 +16,10 @@ from PyQt6.QtWidgets import (
     QGroupBox,
     QHBoxLayout,
     QLabel,
-    QLineEdit,
     QMessageBox,
     QProgressBar,
     QPushButton,
     QScrollArea,
-    QSizePolicy,
-    QSlider,
     QTabWidget,
     QVBoxLayout,
     QWidget,
@@ -53,6 +50,7 @@ from movement_optimizer.rendering import Palette, get_chart_color
 
 from . import plot_renderer
 from .motion_analysis_panel import MotionAnalysisPanel
+from .motion_controls import NumericControl, scrollable_control_panel
 from .policy_trace_canvas import PolicyTraceCanvas, refresh_policy_trace_palette
 from .policy_worker import PolicyOptimizationWorker
 from .vector_overlay import (
@@ -156,99 +154,6 @@ def _chain_overlay_scene(
             vec = (float(field.net_force_n[index][0]), float(field.net_force_n[index][1]))
             arrows.append(ForceArrow(origin, vec, VectorStyle(ARM)))
     return OverlayScene(arrows=tuple(arrows))
-
-
-class NumericControl(QWidget):
-    """Slider plus typed value field without spin-box arrows."""
-
-    valueChanged = pyqtSignal(float)  # noqa: N815 - Qt signal naming convention.
-
-    def __init__(
-        self,
-        lower: float,
-        upper: float,
-        value: float,
-        *,
-        integer: bool = False,
-        decimals: int = 3,
-        steps: int = 1000,
-    ) -> None:
-        super().__init__()
-        if upper <= lower:
-            raise ValueError("upper must be greater than lower")
-        self._lower = lower
-        self._upper = upper
-        self._integer = integer
-        self._decimals = 0 if integer else decimals
-        self._steps = max(1, int(upper - lower) if integer else steps)
-        self._value = self._coerce(value)
-
-        layout = QHBoxLayout(self)
-        layout.setContentsMargins(0, 0, 0, 0)
-        layout.setSpacing(8)
-        self.setMinimumHeight(32)
-        self.slider = QSlider(Qt.Orientation.Horizontal)
-        self.slider.setRange(0, self._steps)
-        self.slider.setTracking(False)
-        self.slider.setMinimumHeight(28)
-        self.slider.setSizePolicy(QSizePolicy.Policy.Expanding, QSizePolicy.Policy.Fixed)
-        self.edit = QLineEdit()
-        self.edit.setFixedWidth(88)
-        self.edit.setMinimumHeight(28)
-        self.edit.setAlignment(Qt.AlignmentFlag.AlignRight)
-        layout.addWidget(self.slider, stretch=1)
-        layout.addWidget(self.edit)
-
-        self.slider.valueChanged.connect(self._on_slider_changed)
-        self.edit.editingFinished.connect(self._on_text_changed)
-        self._sync_widgets()
-
-    def value(self) -> float:
-        return self._value
-
-    def set_value(self, value: float) -> None:
-        new_value = self._coerce(value)
-        if new_value == self._value:
-            self._sync_widgets()
-            return
-        self._value = new_value
-        self._sync_widgets()
-        self.valueChanged.emit(self._value)
-
-    def _coerce(self, value: float) -> float:
-        bounded = min(max(float(value), self._lower), self._upper)
-        return float(round(bounded)) if self._integer else bounded
-
-    def _value_to_slider(self, value: float) -> int:
-        ratio = (value - self._lower) / (self._upper - self._lower)
-        return round(ratio * self._steps)
-
-    def _slider_to_value(self, position: int) -> float:
-        ratio = position / self._steps
-        return self._coerce(self._lower + ratio * (self._upper - self._lower))
-
-    def _sync_widgets(self) -> None:
-        slider_value = self._value_to_slider(self._value)
-        if self.slider.value() != slider_value:
-            self.slider.blockSignals(True)
-            self.slider.setValue(slider_value)
-            self.slider.blockSignals(False)
-        text = f"{int(self._value)}" if self._integer else f"{self._value:.{self._decimals}f}"
-        if self.edit.text() != text:
-            self.edit.setText(text)
-
-    def _on_slider_changed(self, position: int) -> None:
-        self._value = self._slider_to_value(position)
-        self._sync_widgets()
-        self.valueChanged.emit(self._value)
-
-    def _on_text_changed(self) -> None:
-        try:
-            parsed = float(self.edit.text())
-        except ValueError:
-            self._sync_widgets()
-            return
-        self.set_value(parsed)
 
 
 class MotionCanvas(QWidget):
@@ -416,16 +321,6 @@ class MotionCanvas(QWidget):
             )
 
 
-def _scrollable_control_panel(panel: QWidget) -> QScrollArea:
-    scroll_area = QScrollArea()
-    scroll_area.setWidget(panel)
-    scroll_area.setWidgetResizable(True)
-    scroll_area.setHorizontalScrollBarPolicy(Qt.ScrollBarPolicy.ScrollBarAlwaysOff)
-    scroll_area.setMinimumWidth(340)
-    scroll_area.setSizePolicy(QSizePolicy.Policy.Fixed, QSizePolicy.Policy.Expanding)
-    return scroll_area
-
-
 class _MotionViewMixin:
     """Shared Animation/Plots subtab scaffolding for motion-analysis tabs.
 
@@ -569,7 +464,7 @@ class SwingsetTab(_MotionViewMixin, QWidget):
         control_layout.addWidget(self._build_policy_group())
         control_layout.addWidget(self._build_policy_telemetry_group())
         control_layout.addStretch()
-        self._control_scroll = _scrollable_control_panel(control_panel)
+        self._control_scroll = scrollable_control_panel(control_panel)
         right_layout.addWidget(self._control_scroll)
         layout.addWidget(self._control_panel_widget, 0, 1, 2, 1)
         layout.addWidget(self.metric_label, 2, 0, 1, 2)
